backend/foodgram_project_backend/requirements.txt
asgiref==3.6.0
Django==3.2.18
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
Pillow==9.4.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
django-extensions==3.2.1
flake8==5.0.4
pep8-naming==0.13.3
flake8-broken-line==0.6.0
flake8-return==1.2.0
flake8-isort==6.0.0